refactor(cache): move queryFilterCachable out of permissions-cachable
It gates the ad-hoc query filter, not permissions, and permissions-cachable.ts is
upstream core (#24543) — a fork symbol there just adds rebase noise. New
query-filter-cachable.ts reuses the upstream `filter_has_now` helper; tests move
alongside.
