+++ api/src/utils/query-filter-cachable.test.ts
@@ -1,0 +1,25 @@
+import { expect, test } from 'vitest';
+import { queryFilterCachable } from './query-filter-cachable.js';
+
+test('query filter with $NOW is not cachable', () => {
+	expect(queryFilterCachable({ created_on: { _gt: '$NOW' } })).toBe(false);
+
+	expect(
+		queryFilterCachable({ _and: [{ created_on: { _gt: '$NOW(-1 year)' } }] }),
+	).toBe(false);
+
+	expect(
+		queryFilterCachable({ _or: [{ nested: { some: { _gt: '$NOW' } } }] }),
+	).toBe(false);
+});
+
+test('query filter without $NOW is cachable', () => {
+	expect(queryFilterCachable({ created_on: { _gt: '2021-01-01' } })).toBe(true);
+
+	expect(
+		queryFilterCachable({ _and: [{ created_on: { _gt: '2021-01-01' } }] }),
+	).toBe(true);
+
+	expect(queryFilterCachable(null)).toBe(true);
+	expect(queryFilterCachable(undefined)).toBe(true);
+});
